Modified Logic.Js and deleted all DOM manipulation on it
