+++ src/views/javascripts/_taskFrm.js
@@ -1,7 +1,84 @@
-import {
-  ValidateTaskFrm, Task, closeTaskFrm,
-} from './logic';
+import { Task } from './logic';
 import '../stylesheets/_taskFrm.scss';
+
+const ValidateTaskFrm = (e) => {
+  let listName = document.getElementsByTagName('main');
+  listName = listName[0].id;
+  let localArr = [];
+  if (localStorage.getItem(listName) !== '') {
+    localArr = JSON.parse(localStorage.getItem(listName));
+  }
+  const task = Task();
+
+  let boolean = true;
+  const frm = document.getElementById('new-task');
+  const inputs = frm.querySelectorAll('input, textarea, date, time, select');
+  let div = document.getElementById('error-msg');
+
+  if (div === null) {
+    div = document.createElement('div');
+    div.className = 'error-msg';
+    div.id = 'error-msg';
+  } else {
+    div.innerHTML = '';
+  }
+  inputs.forEach(element => {
+    const { value, name } = element;
+    if (value === '') {
+      const span = document.createElement('span');
+      span.innerText = `${name} can't be empty`;
+      div.appendChild(span);
+      boolean = false;
+    } else {
+      task[name] = value;
+      if (name === 'priority') {
+        task[name] = value;
+      } else if (name === 'date') {
+        const newValue = value.split('-');
+        task[name] = `${newValue[1]}/${newValue[2]}/${newValue[0]}`;
+      } else if (name === 'time') {
+        const newValue = value.split(':');
+        if (newValue[0] === '00') {
+          task[name] = `${12}:${newValue[1]}:00 AM`;
+        } else if (newValue[0] === '12') {
+          task[name] = `${12}:${newValue[1]}:00 PM`;
+        } else if (newValue[0] < 12) {
+          task[name] = `${newValue[0]}:${newValue[1]}:00 AM`;
+        } else if (newValue[0] > 12) {
+          task[name] = `${newValue[0] - 12}:${newValue[1]}:00 PM`;
+        }
+      }
+    }
+  });
+
+  if (e.id === undefined) {
+    task.id = localArr.length;
+  } else {
+    task.id = e.id;
+  }
+
+  if (boolean === true) {
+    if (task.id === e.id) {
+      localArr.forEach((element, index) => {
+        if (element.id === task.id) {
+          localArr[index] = task;
+        }
+      });
+    } else {
+      localArr.push(task);
+    }
+    localStorage.setItem(listName, JSON.stringify(localArr));
+  } else {
+    frm.insertBefore(div, frm.childNodes[1]);
+  }
+  return boolean;
+};
+
+function closeTaskFrm() {
+  const main = document.getElementById('mainContainer');
+  const container = document.getElementsByClassName('task-popup');
+  main.removeChild(container[0]);
+}
 
 const TaskForm = (e) => {
   const mainContainer = document.createElement('div');
